[Vapor 3] Client extensions
Moved some helper methods to a Client extensions to facilitate reuse
